Updates for fall 2019
Ok Ron I emailed you the new schedule that needs to be updated for Riverside.
If you could take down Wickliffe’s schedule and maybe we could just put on there eight hours of in car driving for online students $325.
Must have enrollment certificate from online course in order to start in cars.
Under services you have to change the price for teen remedial classes to $100.
Please also add that they need to bring their court papers with them to that class.
Also on the homepage but he hast to be changed for in cars for the online students $325.